+++ src/main/main.js
@@ -14,7 +14,7 @@ function createWindow() {
     webPreferences: {
       nodeIntegration: false,
       contextIsolation: true,
-      preload: path.join(__dirname, 'preload.js')
+      preload: path.join(__dirname, '../preload.js')
     },
     icon: path.join(__dirname, '../../public/icon.png'),
     show: false,
