+++ src/renderer/TradingRiskCalculator.jsx
