+++ src/shared/calculations.js
